api/requirements.txt
requests
pymongo
pandas
fastapi
uvicorn
psycopg2-binary